add SwipeSlider for DateTimeText
Implemented the function that slide dateTimeText with swipe .

TODO 2 items and swap => 3 items; prev, current, next
TODO item rotation

+++ Assets/Scripts/TodoText.cs
@@ -109,6 +109,10 @@ public class TodoText : MonoBehaviour {
 			return;
 		}
 
+		if (textHistory == null) {
+			return;
+		}
+
 		// En route
 		if (textHistory.Count != currentHistoryIndex + 1) {
 			textHistory.RemoveRange (
